+++ src/main/java/de/uib/configed/Configed.java
@@ -49,10 +49,10 @@ public final class Configed {
 	private static Properties extraLocalization;
 	private static boolean showLocalizationStrings;
 
-	private static String host;
 	private static String user;
 	private static String password;
 	private static String otp;
+	private static boolean useSSO;
 
 	private static boolean optionCLIQuerySearch;
 	private static String savedSearch;
@@ -73,9 +73,10 @@ public final class Configed {
 	private static String paramUser;
 	private static String paramPassword;
 	private static String paramOTP;
+	private static Boolean paramSSO;
 
-	private Configed(String paramHost, String paramUser, String paramPassword, String paramOTP) {
-		setParamValues(paramHost, paramUser, paramPassword, paramOTP);
+	private Configed(String paramHost, String paramUser, String paramPassword, String paramOTP, boolean paramSSO) {
+		setParamValues(paramHost, paramUser, paramPassword, paramOTP, paramSSO);
 
 		Logging.debug("starting ", getClass().getName());
 		Logging.debug("default charset is ", Charset.defaultCharset().displayName());
@@ -103,16 +104,18 @@ public final class Configed {
 		FOpsiLicenseMissingText.reset();
 		LicensingInfoMap.requestRefresh();
 
-		ConfigedMain configedMain = new ConfigedMain(paramHost, paramUser, paramPassword, paramOTP);
+		ConfigedMain configedMain = new ConfigedMain(paramHost, paramUser, paramPassword, paramOTP, paramSSO);
 
 		SwingUtilities.invokeLater(configedMain::init);
 	}
 
-	private static void setParamValues(String paramHost, String paramUser, String paramPassword, String paramOTP) {
+	private static void setParamValues(String paramHost, String paramUser, String paramPassword, String paramOTP,
+			Boolean paramSSO) {
 		Configed.paramHost = paramHost;
 		Configed.paramUser = paramUser;
 		Configed.paramPassword = paramPassword;
 		Configed.paramOTP = paramOTP;
+		Configed.paramSSO = paramSSO;
 	}
 
 	public static String getResourceValue(String key) {
@@ -147,8 +150,8 @@ public final class Configed {
 	}
 
 	private static void addMissingArgs() {
-		if (host == null) {
-			host = Utils.getCLIParam("Host: ");
+		if (ConfigedMain.getHost() == null) {
+			ConfigedMain.setHost(Utils.getCLIParam("Host: "));
 		}
 		if (user == null) {
 			user = Utils.getCLIParam("User: ").toLowerCase(Locale.ROOT);
@@ -163,7 +166,7 @@ public final class Configed {
 
 	private static void processLoginOptions(CommandLine cmd) {
 		if (cmd.hasOption("h")) {
-			host = cmd.getOptionValue("h");
+			ConfigedMain.setHost(cmd.getOptionValue("h"));
 		}
 
 		if (cmd.hasOption("u")) {
@@ -176,6 +179,11 @@ public final class Configed {
 
 		if (cmd.hasOption("otp")) {
 			otp = cmd.getOptionValue("otp");
+		}
+
+		if (cmd.hasOption("sso")) {
+			// Single sign-on is not implemented
+			useSSO = true;
 		}
 	}
 
@@ -325,12 +333,13 @@ public final class Configed {
 
 	private static void checkArgsAndStart() {
 		Logging.debug("initiating configed");
+		String host = ConfigedMain.getHost();
 
 		if (optionCLIQuerySearch) {
 			addMissingArgs();
 			initSavedStates();
 			Logging.debug("optionCLIQuerySearch");
-			SavedSearchQuery query = new SavedSearchQuery(host, user, password, otp, savedSearch);
+			SavedSearchQuery query = new SavedSearchQuery(host, user, password, otp, useSSO, savedSearch);
 
 			query.runSearch(true);
 			Main.endApp(Main.NO_ERROR);
@@ -339,7 +348,7 @@ public final class Configed {
 			initSavedStates();
 			Logging.debug("optionCLIDefineGroupBySearch");
 
-			SavedSearchQuery query = new SavedSearchQuery(host, user, password, otp, savedSearch);
+			SavedSearchQuery query = new SavedSearchQuery(host, user, password, otp, useSSO, savedSearch);
 
 			List<String> newGroupMembers = query.runSearch(false);
 
@@ -348,7 +357,7 @@ public final class Configed {
 		} else if (optionCLISwAuditPDF) {
 			Logging.debug("optionCLISwAuditPDF");
 			SwPdfExporter exporter = new SwPdfExporter();
-			exporter.setArgs(host, user, password, otp, clientsFile, outDir);
+			exporter.setArgs(host, user, password, otp, useSSO, clientsFile, outDir);
 			exporter.addMissingArgs();
 			exporter.run();
 
@@ -356,7 +365,7 @@ public final class Configed {
 		} else if (optionCLISwAuditCSV) {
 			Logging.debug("optionCLISwAuditCSV");
 			SWcsvExporter exporter = new SWcsvExporter();
-			exporter.setArgs(host, user, password, otp, clientsFile, outDir);
+			exporter.setArgs(host, user, password, otp, useSSO, clientsFile, outDir);
 			exporter.addMissingArgs();
 			exporter.run();
 
@@ -368,7 +377,7 @@ public final class Configed {
 			initSavedStates();
 
 			OpsiServiceNOMPersistenceController persistenceController = PersistenceControllerFactory
-					.getNewPersistenceController(host, user, password, otp);
+					.getNewPersistenceController(host, user, password, otp, useSSO);
 
 			UserConfigProducing up = new UserConfigProducing(false, host,
 					persistenceController.getHostInfoCollections().getDepotNamesList(),
@@ -385,7 +394,7 @@ public final class Configed {
 			Logging.info("start configed gui since no options for CLI-mode were chosen");
 		}
 
-		new Configed(host, user, password, otp);
+		new Configed(host, user, password, otp, useSSO);
 	}
 
 	public static void initSavedStates() {
@@ -441,11 +450,7 @@ public final class Configed {
 	}
 
 	private static String getSavedStatesDirectoryName(String locationName) {
-		return locationName + File.separator + host.replace(":", "_");
-	}
-
-	public static void setHost(String host) {
-		Configed.host = host;
+		return locationName + File.separator + ConfigedMain.getHost().replace(":", "_");
 	}
 
 	public static SavedStates getSavedStates() {