+++ src/main/java/de/uib/opsicommand/ConnectionHandler.java
@@ -141,11 +141,16 @@ public class ConnectionHandler {
 	 *         unsuccessful connection.
 	 */
 	public HttpsURLConnection establishConnection(boolean doOutput) {
+		return establishConnection(doOutput, false);
+	}
+
+	public HttpsURLConnection establishConnection(boolean doOutput, boolean useInsecure) {
 		if (serviceURL == null) {
 			return null;
 		}
+		Logging.info(this, "establishing connection with ", serviceURL);
 
-		CertificateValidator certValidator = CertificateValidatorFactory.getValidator();
+		CertificateValidator certValidator = CertificateValidatorFactory.getValidator(useInsecure);
 		HttpsURLConnection connection = null;
 
 		try {