+++ src/main/java/de/uib/utils/datapanel/DefaultEditMapPanel.java
@@ -11,8 +11,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
-import java.util.Set;
-import java.util.function.Function;
 
 import javax.swing.JPanel;
 import javax.swing.JPopupMenu;
@@ -21,7 +19,6 @@ import javax.swing.MenuElement;
 import de.uib.configed.Configed;
 import de.uib.configed.type.ConfigOption;
 import de.uib.opsidatamodel.datachanges.UpdateCollection;
-import de.uib.utils.DataChangedObserver;
 import de.uib.utils.logging.Logging;
 
 // works on a map of pairs of type String - List
@@ -166,16 +163,8 @@ public class DefaultEditMapPanel extends JPanel {
 		mapTableModel.setWrite();
 	}
 
-	public List<String> getNames() {
-		return mapTableModel.getKeys();
-	}
-
-	public void setShowToolTip(boolean b) {
-		showToolTip = b;
-	}
-
-	public void registerDataChangedObserver(DataChangedObserver o) {
-		mapTableModel.registerDataChangedObserver(o);
+	public void setShowToolTip(boolean showToolTip) {
+		this.showToolTip = showToolTip;
 	}
 
 	/**
@@ -196,14 +185,6 @@ public class DefaultEditMapPanel extends JPanel {
 	 */
 	public void setUpdateCollection(UpdateCollection updateCollection) {
 		mapTableModel.setUpdateCollection(updateCollection);
-	}
-
-	public void setReadOnlyEntries(Set<String> keys) {
-		mapTableModel.setReadOnlyEntries(keys);
-	}
-
-	public void setEditableFunction(Function<String, Boolean> isEditable) {
-		mapTableModel.setIsEditable(isEditable);
 	}
 
 	protected void logPopupElements() {