+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-## [4.4.0.9] - 2024-30-10
+## [4.4.0.11] - 2024-11-13
+### Changed
+- Improve login dialog usability
+
+## [4.4.0.10] - 2024-11-06
+### Added
+- Add option to log in via Single Sign On (SSO)
+
+## [4.4.0.9] - 2024-10-30
 ### Fixed
 - Correct path for saving certificates when entering port in login dialog
 - Start configed when trusting the certificate once